new header MsgQueue.h
